Refactor Aws4Signer and Aws4UnsignedPayloadSigner to share a common base class
